profile setup admin left

+++ backend/controllers/uploadController.js
@@ -15,7 +15,7 @@ const uploadProfilePic = asyncHandler(async(req,res)=>{
         user.isAdmin = user.isAdmin
         user.profilePic = fileName
         const updatedUser = await user.save()
-        if(delFile.split(".")[1]!==fileName.split(".")[1] && fs.existsSync(filePath)){
+        if(delFile!=="" && delFile.split(".")[1]!==fileName.split(".")[1] && fs.existsSync(filePath)){
             fs.unlinkSync(filePath)
         }
         res.status(200).json({
